consolidate Tag into Chip
