+++ src/core/Chip/index.stories.tsx
@@ -1,9 +1,27 @@
 import { Args, Story } from "@storybook/react";
 import React from "react";
+import Button from "../Button";
 import Chip from "./index";
 
 const Demo = (props: Args): JSX.Element => {
-  return <Chip size="small" label="default small" {...props} />;
+  const { dismissable } = props;
+  const [visible, setVisible] = React.useState(false);
+
+  const handleShowChip = () => setVisible(true);
+  const handleDismissChip = () => setVisible(false);
+
+  if (dismissable) {
+    return (
+      <>
+        <Button variant="contained" onClick={handleShowChip}>
+          Click me!
+        </Button>
+        {visible && <Chip onDelete={handleDismissChip} {...props} />}
+      </>
+    );
+  } else {
+    return <Chip {...props} />;
+  }
 };
 
 export default {
@@ -13,11 +31,22 @@ export default {
 
 const Template: Story = (args) => <Demo {...args} />;
 
+export const LargeSquareChip = Template.bind({});
+
+LargeSquareChip.args = {
+  dismissable: true,
+  label: "Chip",
+  size: "large",
+  variant: "square",
+};
+
 export const Success = Template.bind({});
 
 Success.args = {
   label: "success",
   status: "success",
+  size: "small",
+  variant: "rounded",
 };
 
 export const Error = Template.bind({});
@@ -25,6 +54,8 @@ export const Error = Template.bind({});
 Error.args = {
   label: "error",
   status: "error",
+  size: "small",
+  variant: "rounded",
 };
 
 export const Warning = Template.bind({});
@@ -32,6 +63,8 @@ export const Warning = Template.bind({});
 Warning.args = {
   label: "warning",
   status: "warning",
+  size: "small",
+  variant: "rounded",
 };
 
 export const Info = Template.bind({});
@@ -39,6 +72,8 @@ export const Info = Template.bind({});
 Info.args = {
   label: "info",
   status: "info",
+  size: "small",
+  variant: "rounded",
 };
 
 export const Pending = Template.bind({});
@@ -46,6 +81,8 @@ export const Pending = Template.bind({});
 Pending.args = {
   label: "pending",
   status: "pending",
+  size: "small",
+  variant: "rounded",
 };
 
 export const Beta = Template.bind({});
@@ -53,4 +90,6 @@ export const Beta = Template.bind({});
 Beta.args = {
   label: "beta",
   status: "beta",
+  size: "small",
+  variant: "rounded",
 };
