+++ src/core/Chip/index.tsx
@@ -1,12 +1,19 @@
+import { ChipProps as RawChipProps } from "@material-ui/core";
+import { Clear } from "@material-ui/icons";
 import React from "react";
 import { ExtraProps, StyledChip } from "./style";
 
-type ChipProps = ExtraProps;
+type ChipProps = ExtraProps & RawChipProps;
 
 export { ChipProps };
 
-const Chip = (props: ChipProps): JSX.Element => {
-  return <StyledChip {...props} />;
+const Tag = (props: ChipProps): JSX.Element => {
+  const { dismissable } = props;
+  if (dismissable) {
+    return <StyledChip deleteIcon={<Clear />} {...props} />;
+  } else {
+    return <StyledChip {...props} />;
+  }
 };
 
-export default Chip;
+export default Tag;
