+++ src/index.ts
@@ -22,8 +22,6 @@ export { default as MenuItem } from "./core/MenuItem";
 export * from "./core/styles";
 export * from "./core/Tabs";
 export { default as Tabs } from "./core/Tabs";
-export * from "./core/Tag";
-export { default as Tag } from "./core/Tag";
 export * from "./core/Tooltip";
 export { default as Tooltip } from "./core/Tooltip";
 export * from "./core/TooltipTableContent";
